requirements.txt
discord.py==2.3.2
aiohttp>=3.8.1
python-dotenv>=1.0.0
asyncpg~=0.30.0
psutil>=5.9.0,<6.0.0